Cleans slh up Fixes problem with chia price widget where it would not show after phone restart or re-installation of app
