+++ app/src/main/java/ninja/bored/chiapublicaddressmonitor/ChiaFiatConversionWidgetReceiver.kt
@@ -5,7 +5,10 @@ import android.appwidget.AppWidgetProvider
 import android.content.Context
 import android.content.Intent
 import android.util.Log
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import ninja.bored.chiapublicaddressmonitor.helpers.Slh
+import ninja.bored.chiapublicaddressmonitor.model.ChiaWidgetRoomsDatabase
 
 class ChiaFiatConversionWidgetReceiver : AppWidgetProvider() {
 
@@ -34,34 +37,32 @@ class ChiaFiatConversionWidgetReceiver : AppWidgetProvider() {
         Log.d(TAG, "onUpdate")
         context?.let {
             Slh.setupWidgetUpdateWorker(context)
-//            all done by worker now
 
-//            val database = ChiaWidgetRoomsDatabase.getInstance(context)
-//            appWidgetManager?.let {
-//                appWidgetIds?.forEach { appWidgetId ->
-//                    GlobalScope.launch {
-//                        val widgetFiatConversionSettingsDao =
-//                            database.getWidgetFiatConversionSettingsDao()
-//                        val widgetFiatConversionSettings =
-//                            widgetFiatConversionSettingsDao.getByID(appWidgetId)
-//                        widgetFiatConversionSettings?.let {
-//                            RemoteViews(
-//                                context.packageName,
-//                                R.layout.chia_public_address_widget
-//                            ).apply {
-//                                val allViews = this
-//                                Slh.updateFiatWidgetWithSettings(
-//                                    widgetFiatConversionSettings,
-//                                    allViews,
-//                                    context,
-//                                    appWidgetId,
-//                                    appWidgetManager
-//                                )
-//                            }
-//                        }
-//                    }
-//                }
-//            }
+            val database = ChiaWidgetRoomsDatabase.getInstance(context)
+            appWidgetManager?.let {
+                appWidgetIds?.forEach { appWidgetId ->
+                    GlobalScope.launch {
+                        val widgetFiatConversionSettingsDao =
+                            database.getWidgetFiatConversionSettingsDao()
+                        val widgetFiatConversionSettings =
+                            widgetFiatConversionSettingsDao.getByID(appWidgetId)
+                        widgetFiatConversionSettings?.let {
+
+                            val chiaLatestConversionDao = database.getChiaLatestConversionDao()
+                            chiaLatestConversionDao.getLatestForCurrency(
+                                widgetFiatConversionSettings.conversionCurrency
+                            )?.let { chiaLatestConversionFromDb ->
+                                Slh.updateFiatWidgetWithData(
+                                    chiaLatestConversionFromDb,
+                                    context,
+                                    appWidgetManager,
+                                    appWidgetId
+                                )
+                            }
+                        }
+                    }
+                }
+            }
         }
     }
 }
